requirements.txt
Django==1.5.1
MySQL-python==1.2.4
django-tastypie==0.10.0
django-tastypie-swagger==0.1.1


